clean up Redshift DATs
